frentegg-auth: refactor out a more general Client
In preparation for integration into mz.
